[rc]: fix first gop ratecontrol working bad bug
since have little data to estimate qp, qp of frame in first gop is
nearly hor-line. now we add a re-encode mechanism to avoid this bug.

re-encode frame if it meets all the conditions below:
1. gop is the first gop
2. type is p frame
3. target_bis is larger than 1.5 * real_bits
and the qp_init will decrese 3 when re-encode.

TODO: maybe use sse to calculate a proper value instead of 3 or add
a much more suitable condition to start this re-encode process.

Change-Id: I2bea5dd31a3f0348a23b40650b2e11bce32ca2db
Signed-off-by: Jung Zhao <jung.zhao@rock-chips.com>
Signed-off-by: Lin Kesheng <lks@rock-chips.com>