[jpege]: Add no extra info path for compatibility
Some old kernel do not support register file with extra info. So when it
is no necessary to add extra info slots just disable it.

The reason for extra info is to handle buffer offset larger than 4M. Due
to low 10bit of 32bit address register is used for file handle and only
high 22bit can be used to as offset from buffer head. Then the offset is
limited to 4M. When we need a offset larger than 22bit we need to add
extra info slot behind register file to let kernel handle it as a
register value patch.

BTW, this is not a good design but it fix the issue so far.

Change-Id: Id5cb694f72b15958164175e402dcc03b08e898b5
Signed-off-by: Herman Chen <herman.chen@rock-chips.com>