[h265d]: Move slot setup to init stage
H265 decoder setup slot when the first sps is read. It will cause a dead
loop on mpp_dec.cpp on checking unused slot count.

The slot setup on parsing stage is better to adapt to different memory
usage. This is better to be improved in future.

Thanks AreaScout and mo123 for reporting this issue at:
https://github.com/rockchip-linux/mpp/commit/44f0ee03533bafd961d2e271a90ea77ad77e107f#commitcomment-29392434

Change-Id: I3835397f6df6e5e17402b8b43dc3985e46dff832
Signed-off-by: Herman Chen <herman.chen@rock-chips.com>