[mpp_enc]: Add hal prepare function
Add prepare function to hal interface.

NOTE: The hal prepare function is the hal hook on user control.
When user config is changed hal can process some cfg before enc_impl
updated the config through syntax in task.

Also when encoder is in low delay mode the user control should not be
processed if the current frame is not finished.

Change-Id: Ie287e05daff0d139cf04d93d97246fa176d14d53
Signed-off-by: Herman Chen <herman.chen@rock-chips.com>