[mpp_dec]: remove sleep when the output queue is empty
When the output queue is empty, sleep 1ms is added to avoid an
infinite loop, which leads to the problem that it takes a long
time to acquire the lock probability when outputting the frame
queue. Remove the sleep operation inside get_frame, and the APP
will ensure the problem of dead loop calls.

Signed-off-by: Xingwen Fang <fxw@rock-chips.com>
Change-Id: Ibbbaa352cf6a9fd386f7e1f944571658295977e5