[mpp_dec]: Fix the field video pts is abnormal issue
The number of input field video frames to the decoder
is half of the number output frame video frames.
When the sorting pts is turned on, only half of the
pts cache is taken out from the pts linked list.

So when decoding field video, only half as many pts are cached.

Change-Id: If9055480365b776475861add1470d68210b65971
Signed-off-by: Hongjin Li <vic.hong@rock-chips.com>