[mpp_buf_slot]: fix field source display err on fast mode
In fast mode, there is a situation:
The current frame is output, when the top field decoding is completed
and the bot field decoding is not completed, because of the top and bot
field share one MppFrame.

Solution: modify the hal_output of SlotStatus frome ont bit to two bits
counter. Using hal_output indicates whether a whole frame completed.

Signed-off-by: Yandong Lin <yandong.lin@rock-chips.com>
Change-Id: I18429399e9188c848385c40f847333f1b46fd6bb