[h264_dpb]: Fix frm with negative poc is marked err issue
[Scenes]
poc has negative value
nal_reference_idc is not 0

[Issue]
Mark the current frame as error frm, when the poc value is
less than the I frame(such as negative poc).

[Solution]
Add a flag to judge whether the I frame has been output.
In the case of current frame poc is smaller than the I frame poc:
1. The current frame will be marked with an error flag, when
the I frame has been output.
2. The current frame will not be marked with an error flag,
when the I frame has not been output.

Change-Id: I164cea77fce031716645b5c5c992a0bcf737dd4e
Signed-off-by: Hongjin Li <vic.hong@rock-chips.com>