[h264d]: Update pts when a new frame is generated.
In split mode, it is only possible to determine whether the
previous frame's segmentation for decoding and PTS update is
completed when the second packet is received. However, in
cases where the second packet starts with SPS (Sequence
Parameter Set) and PPS (Picture Parameter Set), we only
terminate the segmentation for decoding of the previous frame
without updating the PTS. This results in two consecutive
frames with duplicate PTS values.

Signed-off-by: Rimon Xu <rimon.xu@rock-chips.com>
Change-Id: I1e55a8fee9e18b62e538f4e5a263c6b8ffc2f087