[h265d]: add compatible for the order of vps
There is a certain error stream that vps is behind
sps and pps while it is still legal and can be
decoded. But the error detection code will stop
the parse process and return error.

To compat this stream by removing return operation
then this kind of stream can be supported.

Change-Id: I2f90a5efa5a8c63a6fa278f9d0408603943001b2
Signed-off-by: xueman.ruan <xueman.ruan@rock-chips.com>