fix[av1d_parser]: fix pix_fmt calc issue
Error case:

When switching sources in the app, if the protocols of
the two sources are the same, but one is an 8-bit source
and the other is a 10-bit source.

Since the first source has been configured with pix_fmt
as 8bit, it will not be recalculated after switching,
resulting in an exception.

Change-Id: I92b44b9c397209171d8ef0f7a0aa65310aa8496d
Signed-off-by: Hongjin Li <vic.hong@rock-chips.com>